rust update

+++ src/matrix.rs
@@ -1,6 +1,6 @@
 use vec;
-//use std::num::Zero;
-use std::num::Float;
+use std::num::Zero;
+//use std::num::Float;
 use std::fmt;
 use std::ops::{Mul};
 
@@ -155,7 +155,7 @@ impl Matrix4
 
         let length2 = q.length2();
 
-        if length2.abs() == Float::zero() {
+        if length2.abs() == f64::zero() {
             let mut mat = Matrix4 { data : m};
             mat.identity_set();
             return mat;
